requirements.txt
sentence-transformers
faiss-cpu
transformers
accelerate
torch
PyMuPDF
gradio
